refactor(runner): Extract checks into a separate module
